style(lint): clippy --fix, plus two places where its suggestion was wrong
24 machine-applicable sites: 11 collapsible_if in park.rs's interposers, 3
unnecessary_cast, 2 unnecessary_mut_passed, and one each of ok_expect,
manual_c_str_literals, deref_addrof, items_after_test_module, manual_div_ceil,
iter_cloned_collect, needless_late_init, let_and_return.

Read the park.rs diff by hand as planned, because nesting there encodes the order
in which an interposer decides whether to park. The collapse is the same chain of
preconditions ending in "and parking timed out", and the may_park guard lives
exactly as long as before, so it stands.

Two of clippy's suggestions did not. deref_addrof turned `&*(&raw const MODULE)`
into a plain reference to a mutable static, which is what static_mut_refs denies
-- that expression was written the way it was on purpose. The test now reads
fields through the raw pointer and never forms a reference at all.
manual_c_str_literals left an unsafe block around a plain c"" literal, so
unused_unsafe fired; the block is gone.

Checked the two unnecessary_mut_passed against the headers rather than trusting
the lint: PHP 8.5 declares both zend_hash_get_current_data_ex and
zend_hash_get_current_key_ex with `const HashPosition *pos`, and
zend_hash_internal_pointer_reset_ex -- the one that does move it -- still gets a
mutable borrow. nextest 9/9 green.

Remaining: 93 undocumented_unsafe_blocks, 4 type_complexity, 1 question_mark.

+++ crates/ignis/src/php/embed.rs
@@ -75,7 +75,7 @@ impl Engine {
             }
             sys::php_embed_init(argc, argv_ptrs.as_mut_ptr())
         };
-        if rc != sys::SUCCESS as i32 {
+        if rc != sys::SUCCESS {
             bail!("php_embed_init failed ({rc})");
         }
         #[cfg(php_async_abi)]
